Refactor cleanup of zoom
